Lowercase package name rather than camel-case
Using LinPac for the tarball name caused the package name, and
therefore installation paths, to also user LinPac. That causes a lot of
location changes (from a user's viewpoint) because in previous versions
the paths were all lower case.

The tarball, package name (as a variable), and paths now user lowercase
'linpac'.

configure.ac was modified.