Add support for PBBS message retrieval
Restructured getmsg so that it can support multiple BBS types, based
upon the BBS type as set in the HOME_TYPE entry in init.mac. The
default is FBB, for backwards compatibility. Newly supported is the
PBBS type. All of the BBS-specific functionality has been broken
out into a base BBS class and a subclass per BBS type. The original
functional breakdown is largely the same as before, thus requiring
minimal change to the existing FBB code.