Clean up mismatched allocations and frees
There were many cases in which a string was being allocated using
strdup() but then freed using delete[] instead of free(). In some
cases, allocation of a given string used either strdup() or new[]
based on a condition, but was always freed using delete[]. All
allocations and frees are now matched, such that strings created
with strdup() or malloc() are destroyed with free().