Fix bug in modifying EventGate children
When an event gate is finished, it's removed from the list of children
maintained by the base gate, and the new max socket number is updated.
The vector::erase function was being used incorrectly to do this, such
that a gate was missed, and the new max socket could then be incorrect.
Fixed the usage of ::erase within an iterator to address this.