Add new command for integrated BBS mail retrieval
Add a new external command that retrieves a message index from a BBS
and then retrieves any messages that do not exist locally. The command
is implemented as a shell script that uses axgetlist to retrieve the
message index and invokes the Linpac getmsg command to retrieve any
missing messages. The shell API is used to retrieve the BBS callsign
and to invoke getmsg.