Fix initial scroll position in index view
The initial scroll position was causing the index to show only the
"last read" message at the top of the index, so that the user needs
to scroll up to see other messages in the list. Fix this to place
the last read message as close as possible to the center of the
screen, adjusting as necessary for messages near the top or bottom
of the message list.